Pos+Neg Freq Cpts
Added code to output positive and negative frequency components together in the same file. Now works as expected, feel better that all data is being output. May make change so that freq data is output along with all other data. Also want to check that code can compute inverse FFT. Suspect that this requires positve and negative cpts in order to work correctly. Examples of FFT are complete and plotted.