Tophat and Signum
The issue is related to the no. of samples and the padding. In the case of both functions the padding supplied in the case of Nsmpls not being a POT is adding extra zeroes to the end and this is disrupting the computation of the FFT. Both Signum and Tophat look like different functions when their FFT is computed. Try again with correct Nsmpls being POT.