TwoFFT
have fixed the issue with twofft. It is possible to simultanesouly compute two FFT, but I don't know if it's worth the hassle to get the code to the point where it's perfect. I don't really want to use it anyway. I want to use four1 because it has an unambiguous input and output, this is not the case for realft and twofft.